// types.ts
export enum Status4D {
  RETAIN = 'RETAIN',
  REMOVE = 'REMOVE',
  PROPOSED = 'PROPOSED',
  FUTURE = 'FUTURE',
  MODIFIED = 'MODIFIED'
}

export enum Ownership {
  LUMEN = 'LUMEN',
  CUSTOMER = 'CUSTOMER'
}

// Instances
export interface Rack {
  id: string;
  definitionId: string; // Link to RackDefinition
  label: string; // Typically "LineUp.Bay" e.g. "001.005"
  lineUp: string; // e.g. "001"
  bayNo: string;  // e.g. "005"
  floor: string;  // e.g. "001"
  suite: string;  // e.g. "003"
  location: string; // Full Pro Inventory Address
  x: number;
  y: number;
  rotation?: number;
  status: Status4D;
  ownerId: string;
  virtualTourUrl?: string;
  proInventoryId?: string;
  isVerified?: boolean;
}

export interface Equipment {
  id: string;
  definitionId: string; // Link to EquipmentDefinition
  name: string;
  rackId: string;
  baseRMU: number; // PDF: "RMU 8 is the base"
  status: Status4D;
  ownerId: string;
  vendorId: string;
  model?: string;
  proInventoryId?: string;
  isVerified?: boolean;
  power?: string;
  description?: string;
  datasheetUrl?: string;
}